Published Blazor tracing behaves poorly
The published version runs and traces but does not propagate the trace from Blazor Client to Server because it does not include the traceparent header.
I could add it but I feel like I am just trying to force a ActivityListener to work on a single threaded environment when it shouldn't.  If anything I could write special exporters not dependent  an ActivityListener host.